fix: 훅 컴포넌트 분리
