+++ src/hooks/home/useMedicineList.js
@@ -1,12 +1,14 @@
 import { useState } from 'react';
 
-const useMedicineList = () => {
-  const [medi, setMedi] = useState([
-    { id: 1, text: '오메가3', time: null, completed: false },
-    { id: 2, text: '메가슬립 수면엔 미강테아닌', time: null, completed: false },
-  ]);
+// 약 목록을 관리하는 커스텀 훅
+function useMedicineList(initialMedi) {
+  // 약 목록을 저장하는 상태
+  const [medi, setMedi] = useState(initialMedi);
 
-  // 영양제 체크 상태 변경 함수
+  // 모달의 표시 여부를 저장하는 상태
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  // 약의 완료 상태를 토글하는 함수
   const handleToggle = (id) => {
     const updatedMedi = medi.map((item) =>
       item.id === id ? { ...item, completed: !item.completed } : item
@@ -14,7 +16,17 @@ const useMedicineList = () => {
     setMedi(updatedMedi);
   };
 
-  // 새로운 영양제 추가 함수
+  // 모달을 여는 함수
+  const openModal = () => {
+    setIsModalOpen(true);
+  };
+
+  // 모달을 닫는 함수
+  const closeModal = () => {
+    setIsModalOpen(false);
+  };
+
+  // 새로운 약을 목록에 추가하는 함수
   const addNewMedicine = (newMedicine, time) => {
     const nextId = medi.length + 1;
     const newMedi = [
@@ -24,11 +36,22 @@ const useMedicineList = () => {
     setMedi(newMedi);
   };
 
+  // 약을 목록에서 삭제하는 함수
+  const deleteMedicine = (id) => {
+    const updatedMedi = medi.filter((item) => item.id !== id);
+    setMedi(updatedMedi);
+  };
+
+  // 훅에서 반환하는 값과 함수들
   return {
     medi,
+    isModalOpen,
     handleToggle,
+    openModal,
+    closeModal,
     addNewMedicine,
+    deleteMedicine,
   };
-};
+}
 
 export default useMedicineList;
